tegra-bootinfo: use different offsets on t210 platforms
There is no pseudo-GPT on t210 platforms, and we don't
want our storage block to interfere with the VER/VER_b
blocks, so locate the primary at the very end of the
SPI flash/boot1 device, and the backup copy between
VER_b and VER.  There are 64KiB allocated for each
VER, which goes mostly unused, so these locations
should be relatively safe.

Also bump the default extension size to 15, as we
should have room for 8KiB on all platforms.

Signed-off-by: Matt Madison <matt@madison.systems>