Hard-code BCT page/block sizes on t18x/t19x
The page and block size fields are in the encrypted
section of the BCT on t18x/t19x platforms, so we can't
use them for determining how to update the BCT.
The two sizes vary depending on the boot device (eMMC
or SPI flash), but otherwise don't change, so just
use the known values.

Back-ported from commit cfde4e1.

Signed-off-by: Matt Madison <matt@madison.systems>