smd: fix priority handling when marking boot successful
When marking a boot in the current slot as successful, make
sure that the priorty level of the other slot remains below
that of the current slot.  Turns out that when the current
slot is 1, and the priority of slot 0 gets to 13, the bootloader
resets slot 0's priority to 15, and since the two slots have
the same priority on the next reboot, we'll switch back to slot 0
again, which is not the desired behavior.

So make sure that we force the other slot's priority to one
less than the current slot's priority, to remain in the current
slot on the next reboot.

Signed-off-by: Matt Madison <matt@madison.systems>