tegra-bootinfo: fix --initialize and --force-initialize handling
With the split of the bootinfo low-level code into a library, part
of the initialization logic got slightly mangled.

For the case where initialization is requested and we find no valid
bootinfo blocks, set the 'current' index to 0 after zeroing out both
blocks, rather than leaving it at -1 (which triggers a spurious error).

For the force-init case, add a BOOTINFO_O_FORCE_INIT flag to bootinfo_open().
When present, initialization is done even if we find a valid bootinfo block.

Signed-off-by: Matt Madison <matt@madison.systems>